adding core, management and and literary events page.
however the background has'nt been added yet.

The core events given here, and the non-core events which will be pushed asap b one of our members , will fall under the same category of "technical events". so basically two html pages in a row connected.
And so we have  3 events in total.